Separate webnative module into init and account
